requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langchain==0.1.4
langchain-anthropic==0.1.1
pydantic==2.5.3
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
playwright==1.41.0
pymysql==1.1.0
redis==5.0.1